requirements.txt
openai>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
certifi>=2024.2.2
pytz>=2024.1
playwright>=1.40.0
jmespath>=1.0.1
scrapfly-sdk>=0.8.23